requirements.txt
streamlit
requests
pandas
numpy
scikit-learn
tensorflow
statsmodels
yfinance
newsapi-python
pyportfolioopt
stable-baselines3
matplotlib
shimmy>=2.0
gymnasium
langchain
langchain_community
openai
transformers
huggingface_hub
beautifulsoup4
yahooquery
sec-edgar-downloader
textblob
vaderSentiment
scipy
cvxpy
PyPortfolioOpt
fpdf
riskfolio-lib
feedparser
vaderSentiment
reportlab
groq
faiss-cpu
fpdf
xhtml2pdf
markdown2
tabulate
langchain-groq
sentence-transformers
langchain-community
prophet
tf-keras
chromadb